Feature/notification service

+++ src/main/java/com/study/studypal/notification/dto/notification/response/UnreadNotificationCountResponseDto.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @NoArgsConstructor
 @Builder
 public class UnreadNotificationCountResponseDto {
-  private int count;
+  private long count;
 }
